+++ backend/src/price-is-nice/guessInformation.ts
